Add logic request order view

+++ App/go_quick_app/lib/models/hoa_don.dart
@@ -29,7 +29,7 @@ class HoaDon {
 
   int? maHoaDon;
   NhanVien? nguoiLapHoaDon;
-  int? tongThanhTien;
+  double? tongThanhTien;
   ThanhToan? thanhToan;
   int? soNguoi;
   Ban? ban;
